refactor: change PanGu to lib
AVOID defining a class that contains only static members.
